Migrate Axis client to Axis2
This commit migrates the Airline client to Axis2.

The biggest problem was that Axis2 parses date differently and always
adds a timezone. As a result, the input string is no longer compliant
with ISO 8601. The server has been updated to tolerate both that format
and the one with a timezone attached.

The build is also simpler as it leverages the Axis2 wsdl2code maven
plugin, rather than calling wsdl2java via ant.

Closes gh-16